requirements.txt
numpy
pytest
openai>=1.0.0
pyyaml